Changing SelectionItem to use Integer instead of Short for id, to avoid overflow errors in report params dropdown
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@13229 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/reports/business/AbstractReportParameterForm.java
@@ -42,9 +42,9 @@ public abstract class AbstractReportParameterForm implements
 	public abstract String[] getAllFormParameterNames();
 	
 	protected void addErrorIfInvalid(Errors errors, String input,
-			Short invalidValue, String fieldName, String errorCode) {
+			Integer invalidValue, String fieldName, String errorCode) {
 		if (!NumberUtils.isDigits(input)
-				|| invalidValue.equals(Short.valueOf(input))) {
+				|| invalidValue.equals(Integer.valueOf(input))) {
 			errors.rejectValue(fieldName, errorCode);
 		}
 	}
